Fix: remove right element in useFieldArray remove The core of the issue is due to how React handles child components. Take the following example:
```tsx
<div>
  {[1,2].map((_, index) => (
    <MyInput key={index} />
  ))}
</div>
```

where `MyInput` is a stateful component.

No matter which element (i.e. 1 or 2) we remove of the array the first
rendered element will remain, as it will have index 0 in both cases.
Removing the `key` prop altogether still causes the last element to be
removed.

In the useFieldArray this has the effect that removing any element
actually ends up removing the last element. This isn't a bug in
`useFieldArray` per se (i.e. it does remove the right element from its
internal arrays) but rather an issue of its children (e.g. components
using `useField`) being stateful.

We could try to hack around this somehow by creating some internal,
unique key props to try to convince React to remove the child component
we want but that's brittle.

The key issue here is that all the hooks (e.g. `useField` and `useForm`)
have their own copy of the state (and thus they are all stateful).
`useField` just doesn't care if e.g. `useFieldArray` changes its state
(by removing an element). This is by design as the design is to have
state changes flow "upward" from `useField` towards `useForm`. However
in the case of `remove` we actually need state to also flow "downwards".

The solution here is to simply not have copies of the state and only
have `useForm` own the form state. In that way there are no state in the
`useField` (at least as far as the form value is concerned) and thus
changing the state by e.g. removing an array element will be reflected
correctly in the children.