Simplify parent notification in useFieldArray/useFieldObject
Before we relied in `useRef` and manually calling `onChange` in every
place we changed the ref. Now we instead use `useState` and rely on a
single `useEffect` to notify the parent no matter the reason the state
changed, which means we have 1 place to get correct instead of N places.